refactor: separate concerns into 3-layer architecture (core/CLI/web)
Extract the 2272-line aggregate_collect.py monolith into clean pipeline
modules. Eliminate sys.argv hack and module-level side effects.

- Add scilex/config.py: centralized SciLExConfig with from_files() and
  from_dicts() constructors, replacing scattered config loading
- Add scilex/pipeline/ modules: tracker, text_filter, citation_filter,
  ranking, itemtype_filter, enrichment, post_filter, orchestrator
- Add pipeline/orchestrator.py: run_aggregation() and run_collection()
  as pure function calls, accepting SciLExConfig + AggregationOptions
- Replace sys.argv manipulation in scilex_api.py with orchestrator call
- Remove sys.path.insert hacks from web files
- Move module-level side effects (setup_logging, load_all_configs, print)
  inside main() in run_collection.py and aggregate_collect.py
- Centralize FORMAT_CONVERTERS registry in crawlers/aggregate.py
- Standardize logging via setup_logging() in all entry points
- Add shared post_filter.py for web UI/API filtering consolidation
- Slim aggregate_collect.py from 2272 to ~960 lines (thin CLI wrapper)
