+++ scilex/pipeline/__init__.py
@@ -1,0 +1,5 @@
+"""SciLEx pipeline modules.
+
+Extracted from aggregate_collect.py for separation of concerns.
+Each module handles one stage of the aggregation pipeline.
+"""
